requirements.txt
bokeh==2.1.1
future==0.18.2
Jinja2==2.11.2
joblib==0.16.0
llvmlite==0.33.0
MarkupSafe==1.1.1
numba==0.50.1
numpy==1.19.0
packaging==20.4
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyts==0.11.0
pywarm==0.4.1
PyYAML==5.3.1
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
TA-Lib==0.4.18
threadpoolctl==2.1.0
torch==1.5.1
tornado==6.0.4
typing-extensions==3.7.4.2
